requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
langchain>=0.2.0
langgraph>=0.1.0
chromadb>=0.5.0
sentence-transformers>=2.7.0
langchain-google-genai>=1.0.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
pypdf>=4.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
google-generativeai>=0.5.0
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
pytesseract>=0.3.10
Pillow>=10.0.0
